restructuring and partitioning gnutls crypto objects
